WIP - split register and config pages
